fix: python 3.13 and pytorch 2.7 compatibility
- Replace deprecated np.int with np.int32 in action_dict.py

- Fix random.shuffle API change in colors.py (Python 3.13)

- Replace torch._six import with math.inf (PyTorch 2.7)

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ distar/pysc2/lib/action_dict.py
@@ -923,7 +923,7 @@ def merge_judge(target_general_action, val):
 
 GENERAL_ACTION_INFO_MASK = {}
 ACT_TO_GENERAL_ACT = {}
-ACT_TO_GENERAL_ACT_ARRAY = np.full(max(ACTION_INFO_MASK.keys()) + 1, -1, dtype=np.int)
+ACT_TO_GENERAL_ACT_ARRAY = np.full(max(ACTION_INFO_MASK.keys()) + 1, -1, dtype=np.int32)
 for k, v in ACTION_INFO_MASK.items():
     general_id = v['general_id']
     if general_id is None or general_id == 0:
